itc: address correctness-review findings (M1-M3, L1-L2)
M1: drive the bounded lazy-skip at scale for is_disjoint/contains/leq via
    misaligned pairs (a shallow leaf aligned against a deep subtree at every
    level => Theta(n) distinct bounded skips), instead of self-comparison
    where leaves always aligned with leaves. Bite-checked: a reintroduced
    unbounded skip fails the guard at ~16x.
M2: add a Bushy id shape + grow_bushy_is_linear so grow's multi-region cost
    comparison (cl<cr with neither child COST_MAX) is exercised at scale.
M3: count unpack/repack in step! so the tick complexity test covers the full
    transcode pipeline, not just fill/grow.
L1: thread leq offsets as u64 (matching the event ops, the oracle, and the
    stored bases); document why no real tree can overflow.
L2: debug_assert the chosen-path id leaf is full in grow_emit, naming the
    never-on-path-empty-region invariant.

+++ crates/itc/src/version/compare.rs
@@ -9,8 +9,11 @@
 //! walk reports where each subtree ended, so a sibling resumes there instead of
 //! skipping the left subtree. The one place `b` is skipped is under an `a` leaf, which
 //! dominates `b`'s whole subtree there; each `b` node is skipped at most once, so the
-//! total stays linear. Path sums are threaded as `u128` offsets, which cannot overflow
-//! for any in-memory tree.
+//! total stays linear. Path sums are threaded as `u64` offsets — the same width as the
+//! stored bases and as `ev_join`/`fill`/`grow` and the oracle. They cannot overflow for
+//! any real tree: a path sum is the running total of stored bases along a root-to-node
+//! path, `tick` adds 1 at a time, and any tree small enough to hold in memory has a
+//! total event count far below `u64::MAX`.
 
 use core::cmp::Ordering;
 
@@ -60,16 +63,16 @@ enum Job {
     /// Compare the subtrees at these positions, under these path-sum offsets.
     Eval {
         ap: usize,
-        ao: u128,
+        ao: u64,
         bp: usize,
-        bo: u128,
+        bo: u64,
     },
     /// Both-internal node: its left child finished; launch the right child, whose `b`
     /// position is threaded from where the left child's `b` subtree ended.
-    RightLockstep { an: u128, bn: u128 },
+    RightLockstep { an: u64, bn: u64 },
     /// Broadcast node (`b` is a leaf): launch the right child against the same pinned
     /// `b` leaf, with `a`'s position threaded from the left child's end.
-    RightBroadcast { an: u128, bp: usize, bo: u128 },
+    RightBroadcast { an: u64, bp: usize, bo: u64 },
 }
 
 /// Whether `a`'s event function is pointwise `<=` `b`'s. Iterative and `O(n + m)`.
@@ -88,8 +91,8 @@ fn leq(a: &EvView, b: &EvView) -> bool {
             Job::Eval { ap, ao, bp, bo } => {
                 let (a_internal, a_base, a_next) = a.header(ap);
                 let (b_internal, b_base, b_next) = b.header(bp);
-                let an = ao + a_base as u128;
-                let bn = bo + b_base as u128;
+                let an = ao + a_base;
+                let bn = bo + b_base;
                 if an > bn {
                     return false;
                 }
